Remove white-space: normal; from label
I've got a toggle button in a dropdown, and don't want the button's label to be wrapped if a scroll bar is displayed on the dropdown. However, I can't do this by styling the toggle button with `white-space: nowrap;`, because the label is styled with `white-space: normal;`. Since normal is the default, this commit removes it to allow clients to style the text content how they wish.